@@ -1,23 +1,23 @@
 {
     "name": "orkg-widget",
     "scripts": {
-        "build": "./node_modules/.bin/webpack --config webpack.config.js",
-        "build-release": "./node_modules/.bin/webpack --config webpack.config.js --env.prod",
-        "start": "./node_modules/.bin/webpack-dev-server --open"
+        "build": "./node_modules/.bin/webpack --config webpack.config.js --mode=development",
+        "build-release": "./node_modules/.bin/webpack --config webpack.config.js --mode=production",
+        "start": "./node_modules/.bin/webpack serve --port=9060 --mode=development"
     },
     "devDependencies": {
-        "@babel/core": "^7.0.0-beta.32",
-        "@babel/preset-env": "^7.0.0-beta.32",
-        "babel-loader": "^8.0.0-beta.0",
-        "copy-webpack-plugin": "^4.2.0",
-        "css-loader": "^3.2.0",
-        "html-loader": "^0.5.1",
-        "style-loader": "^0.19.0",
-        "webpack": "^3.8.1",
-        "webpack-dev-server": "^2.9.4"
+        "@babel/core": "^7.12.10",
+        "@babel/preset-env": "^7.12.11",
+        "babel-loader": "^8.2.2",
+        "copy-webpack-plugin": "^7.0.0",
+        "css-loader": "^5.0.1",
+        "file-loader": "^6.2.0",
+        "html-loader": "^1.3.2",
+        "style-loader": "^2.0.0",
+        "terser-webpack-plugin": "^5.0.3",
+        "url-loader": "^4.1.1",
+        "webpack": "^5.11.1",
+        "webpack-cli": "^4.3.0"
     },
-    "dependencies": {
-        "file-loader": "^5.0.2",
-        "url-loader": "^3.0.0"
-    }
+    "dependencies": {}
 }
