@@ -1,94 +1,107 @@
-import React, { Component } from 'react';
-import { ListGroup, ListGroupItem } from 'reactstrap';
+import React, { useState, useEffect } from 'react';
+import { ListGroup, ListGroupItem, Button, ButtonDropdown, DropdownToggle, DropdownMenu, DropdownItem } from 'reactstrap';
 import { Link } from 'react-router-dom';
 import ROUTES from 'constants/routes.js';
 import { getStatementsBySubjects } from 'services/backend/statements';
 import { getResourcesByClass } from 'services/backend/resources';
-import { getPaperData, sortMethod } from 'utils';
-import { find } from 'lodash';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
 import { reverse } from 'named-urls';
+import { getPaperData } from 'utils';
+import { find } from 'lodash';
 import { CLASSES } from 'constants/graphSettings';
 
-class FeaturedPapers extends Component {
-    state = {
-        papers: null
-    };
+export default function FeaturedPapers() {
+    const [papers, setPapers] = useState([]);
+    const [isLoading, setIsLoading] = useState(false);
+    const [filter, setFilter] = useState('featured');
+    const [dropdownOpen, setOpen] = useState(false);
 
-    componentDidMount() {
-        this.loadfeaturedPapers();
-    }
+    const toggle = () => setOpen(!dropdownOpen);
 
-    loadfeaturedPapers = async () => {
-        this.setState({ loadingPapers: true });
-        getResourcesByClass({
-            id: CLASSES.FEATURED_PAPER,
-            page: 1,
-            items: 7,
-            desc: true
-        }).then(result => {
-            getStatementsBySubjects({ ids: result.map(p => p.id) })
-                .then(papersStatements => {
-                    let papers = papersStatements.map(paperStatements => {
-                        const paperSubject = find(result, { id: paperStatements.id });
-                        return getPaperData(
-                            paperStatements.id,
-                            paperSubject && paperSubject.label ? paperSubject.label : 'No Title',
-                            paperStatements.statements
-                        );
-                    });
-
-                    // order featured comparison
-                    papers = papers.sort((p1, p2) => sortMethod(p1.order, p2.order));
-
-                    this.setState({
-                        papers: papers
-                    });
-                })
-                .catch(error => {
-                    this.setState({
-                        papers: null
-                    });
-                    console.log(error);
-                });
+    const loadPapers = async paperFilter => {
+        setIsLoading(true);
+        const responseJson = await getResourcesByClass({
+            id: paperFilter === 'featured' ? CLASSES.FEATURED_PAPER : CLASSES.PAPER,
+            sortBy: 'created_at',
+            desc: false,
+            items: 10
         });
+        const ids = responseJson.map(paper => paper.id);
+        getStatementsBySubjects({
+            ids
+        })
+            .then(papersStatements => {
+                const papers = papersStatements.map(paperStatements => {
+                    const resourceSubject = find(responseJson, { id: paperStatements.id });
+                    return getPaperData(
+                        paperStatements.id,
+                        paperStatements && resourceSubject.label ? resourceSubject.label : 'No Title',
+                        paperStatements.statements
+                    );
+                });
+
+                setPapers(papers);
+                setIsLoading(false);
+            })
+            .catch(() => {
+                setPapers([]);
+                setIsLoading(false);
+            });
     };
 
-    render() {
-        return (
-            <div className="pl-3 pr-3 p-4">
-                {this.state.papers ? (
-                    this.state.papers.length > 0 ? (
-                        <>
-                            <ListGroup>
-                                {this.state.papers.map((paper, index) => (
-                                    <ListGroupItem key={index} className="p-0 m-0 mb-4" style={{ border: 0 }}>
-                                        <h5 className="h6">
-                                            <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: paper.id })} style={{ color: 'inherit' }}>
-                                                {paper.label ? paper.label : <em>No title</em>}
-                                            </Link>
-                                        </h5>
-                                        {paper.authorNames && paper.authorNames.length > 0 && (
-                                            <span className="badge badge-lightblue"> {paper.authorNames[0].label}</span>
-                                        )}
+    useEffect(() => {
+        loadPapers(filter);
+    }, [filter]);
 
-                                        {/*<span className="badge badge-lightblue"> {paper.paperItem.researchField}</span> // reserach fields can be long which doesn't look like in a badge here*/}
-                                    </ListGroupItem>
-                                ))}
-                            </ListGroup>
-                        </>
-                    ) : (
-                        <div className="text-center">No papers found</div>
-                    )
-                ) : (
-                    <div className="text-center">
-                        <Icon icon={faSpinner} spin /> Loading
-                    </div>
-                )}
+    return (
+        <div className="pl-3 pr-3 pt-2 pb-3">
+            <div className="d-flex justify-content-end mb-2">
+                <ButtonDropdown size="sm" isOpen={dropdownOpen} toggle={toggle}>
+                    <DropdownToggle caret color="lightblue">
+                        {filter === 'featured' && 'Featured'}
+                        {filter === 'latest' && 'Latest'}
+                    </DropdownToggle>
+                    <DropdownMenu right>
+                        <DropdownItem onClick={() => setFilter('featured')}>Featured</DropdownItem>
+                        <DropdownItem onClick={() => setFilter('latest')}>Latest</DropdownItem>
+                    </DropdownMenu>
+                </ButtonDropdown>
             </div>
-        );
-    }
-}
+            {!isLoading ? (
+                papers.length > 0 ? (
+                    <>
+                        <ListGroup>
+                            {papers.map((paper, index) => (
+                                <ListGroupItem key={index} className="p-0 m-0 mb-4" style={{ border: 0 }}>
+                                    <h5 className="h6">
+                                        <Link to={reverse(ROUTES.VIEW_PAPER, { resourceId: paper.id })} style={{ color: 'inherit' }}>
+                                            {paper.label ? paper.label : <em>No title</em>}
+                                        </Link>
+                                    </h5>
+                                    {paper.authorNames && paper.authorNames.length > 0 && (
+                                        <span className="badge badge-lightblue"> {paper.authorNames[0].label}</span>
+                                    )}
+                                </ListGroupItem>
+                            ))}
+                        </ListGroup>
 
-export default FeaturedPapers;
+                        <div className="text-center">
+                            <Link to={ROUTES.PAPERS}>
+                                <Button color="primary" className="mr-3" size="sm">
+                                    View more
+                                </Button>
+                            </Link>
+                        </div>
+                    </>
+                ) : (
+                    <div className="text-center">{filter === 'featured' ? 'No featured paper found' : 'No paper found'}</div>
+                )
+            ) : (
+                <div className="text-center">
+                    <Icon icon={faSpinner} spin /> Loading
+                </div>
+            )}
+        </div>
+    );
+}
