@@ -1,21 +1,16 @@
-import React, { Component } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Button, Card, ListGroup, ListGroupItem, CardDeck } from 'reactstrap';
-import { getStatementsBySubject } from 'services/backend/statements';
+import { getStatementsBySubjectAndPredicate, getParentResearchFields } from 'services/backend/statements';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
-import { faAngleDoubleRight } from '@fortawesome/free-solid-svg-icons';
-import { connect } from 'react-redux';
-import { compose } from 'redux';
-import { updateResearchField, nextStep, previousStep, openTour, closeTour } from 'actions/addPaper';
+import { faAngleDoubleRight, faSpinner } from '@fortawesome/free-solid-svg-icons';
+import { resourcesUrl } from 'services/backend/resources';
+import Autocomplete from 'components/Autocomplete/Autocomplete';
+import { CLASSES, PREDICATES, MISC } from 'constants/graphSettings';
+import { useSelector, useDispatch } from 'react-redux';
+import { updateResearchField, nextStep, previousStep } from 'actions/addPaper';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
-import styled, { withTheme } from 'styled-components';
-import { withCookies, Cookies } from 'react-cookie';
-import PropTypes from 'prop-types';
-import Tour from 'reactour';
-import Tooltip from '../../Utils/Tooltip';
+import styled from 'styled-components';
 import flattenDeep from 'lodash/flattenDeep';
-import { disableBodyScroll, enableBodyScroll, clearAllBodyScrollLocks } from 'body-scroll-lock';
-import { MISC } from 'constants/graphSettings';
-import env from '@beam-australia/react-env';
 
 const ListGroupItemTransition = styled(CSSTransition)`
     &.fadeIn-enter,
@@ -47,58 +42,83 @@ const FieldSelector = styled(Card)`
  * Component for selecting the research field of a paper
  * This might be redundant in the future, if the research field can be derived from the research problem
  */
+function ResearchField() {
+    const [showError, setShowError] = useState(false);
+    const [isLoading, setIsLoading] = useState(false);
+    const selectedResearchField = useSelector(state => state.addPaper.selectedResearchField);
+    const researchFields = useSelector(state => state.addPaper.researchFields);
 
-class ResearchField extends Component {
-    constructor(props) {
-        super(props);
+    const dispatch = useDispatch();
 
-        this.state = {
-            showError: false
-        };
-
-        // check if a cookie of take a tour exist
-        if (this.props.cookies && this.props.cookies.get('taketour') === 'take' && !this.props.cookies.get('showedReaseachFiled')) {
-            this.props.openTour();
-            this.props.cookies.set('showedReaseachFiled', true, { path: env('PUBLIC_URL'), maxAge: 604800 });
-        }
-    }
-
-    componentDidMount() {
+    useEffect(() => {
         // select the main field is none is selected yet (i.e. first time visiting this step)
-        if (this.props.selectedResearchField === '') {
-            this.getFields(MISC.RESEARCH_FIELD_MAIN, 0);
+        if (selectedResearchField === '') {
+            getFields(MISC.RESEARCH_FIELD_MAIN, 0);
         }
-    }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
 
-    componentWillUnmount() {
-        clearAllBodyScrollLocks();
-    }
-
-    disableBody = target =>
-        disableBodyScroll(target, {
-            reserveScrollBarGap: true
-        });
-    enableBody = target => enableBodyScroll(target);
-
-    handleNextClick = () => {
-        // TODO validation: check if a research field is selected
-        if (this.props.selectedResearchField === MISC.RESEARCH_FIELD_MAIN) {
-            this.setState({
-                showError: true
-            });
-
+    const handleNextClick = () => {
+        if (selectedResearchField === MISC.RESEARCH_FIELD_MAIN) {
+            setShowError(true);
             return;
         }
-
-        this.props.nextStep();
+        dispatch(nextStep());
     };
 
-    getFields(fieldId, level) {
-        getStatementsBySubject({ id: fieldId }).then(res => {
-            let researchFields = [];
+    const handleFieldClick = (fieldId, currentLevel) => {
+        getFields(fieldId, currentLevel + 1);
+    };
+
+    const handleFieldSelect = selected => {
+        setIsLoading(true);
+        // current researchFields
+        getParentResearchFields(selected.id).then(parents => {
+            parents = parents.reverse();
+            Promise.all([
+                researchFields[0].map(elm => ({
+                    ...elm,
+                    active: elm.id === parents[1]?.id ? true : false
+                })),
+                ...parents.slice(1).map((parent, i) =>
+                    getStatementsBySubjectAndPredicate({
+                        subjectId: parent.id,
+                        predicateId: PREDICATES.HAS_SUB_RESEARCH_FIELD
+                    }).then(res =>
+                        res
+                            .map(elm => ({
+                                label: elm.object.label,
+                                id: elm.object.id,
+                                active: elm.object.id === parents[i + 2]?.id ? true : false
+                            }))
+                            .sort((a, b) => {
+                                // sort research fields alphabetically
+                                return a.label.localeCompare(b.label);
+                            })
+                    )
+                )
+            ]).then(data => {
+                dispatch(
+                    updateResearchField({
+                        researchFields: data,
+                        selectedResearchField: selected.id
+                    })
+                );
+                setIsLoading(false);
+            });
+        });
+    };
+
+    const getFields = (fieldId, level) => {
+        setIsLoading(true);
+        getStatementsBySubjectAndPredicate({
+            subjectId: fieldId,
+            predicateId: PREDICATES.HAS_SUB_RESEARCH_FIELD
+        }).then(res => {
+            let subResearchFields = [];
 
             res.forEach(elm => {
-                researchFields.push({
+                subResearchFields.push({
                     label: elm.object.label,
                     id: elm.object.id,
                     active: false
@@ -106,12 +126,11 @@ class ResearchField extends Component {
             });
 
             // sort research fields alphabetically
-            researchFields = researchFields.sort((a, b) => {
+            subResearchFields = subResearchFields.sort((a, b) => {
                 return a.label.localeCompare(b.label);
             });
-
-            const researchFieldsNew = [...this.props.researchFields];
-            researchFieldsNew[level] = researchFields;
+            let researchFieldsNew = [...researchFields];
+            researchFieldsNew[level] = subResearchFields;
 
             // add active to selected field
             if (researchFieldsNew[level - 1] !== undefined) {
@@ -121,182 +140,100 @@ class ResearchField extends Component {
             }
 
             // hide all higher level research fields (in case a lower level research field has been selected again)
-            researchFieldsNew.forEach((elm, index) => {
-                if (index > level) {
-                    delete researchFieldsNew[index];
-                }
-            });
-
-            this.props.updateResearchField({
-                researchFields: researchFieldsNew,
-                selectedResearchField: fieldId
-            });
+            researchFieldsNew = researchFieldsNew.filter((elm, index) => index <= level);
+            setIsLoading(false);
+            return dispatch(
+                updateResearchField({
+                    researchFields: researchFieldsNew,
+                    selectedResearchField: fieldId
+                })
+            );
         });
-    }
-
-    handleFieldClick(fieldId, currentLevel) {
-        if (this.props.isTourOpen) {
-            this.requestCloseTour();
-        }
-        this.getFields(fieldId, currentLevel + 1);
-    }
-
-    requestCloseTour = () => {
-        this.enableBody();
-        this.props.closeTour();
     };
 
-    handleLearnMore = step => {
-        this.props.openTour(step);
-    };
+    let researchFieldLabel;
+    if (researchFields && researchFields.length > 0) {
+        const rF = flattenDeep(researchFields).find(rf => rf.id === selectedResearchField);
+        researchFieldLabel = rF ? rF.label : selectedResearchField;
+    }
 
-    render() {
-        let errorMessageClasses = 'text-danger mt-2 pl-2';
-        errorMessageClasses += !this.state.showError ? ' d-none' : '';
-        let researchFieldLabel;
-        if (this.props.researchFields && this.props.researchFields.length > 0) {
-            const rF = flattenDeep(this.props.researchFields).find(rf => rf.id === this.props.selectedResearchField);
-            researchFieldLabel = rF ? rF.label : this.props.selectedResearchField;
-        }
-
-        return (
-            <div>
-                <h2 className="h4 mt-4 mb-5">
-                    <Tooltip
-                        message={
-                            <span>
-                                Select the most appropriate research field for the paper.{' '}
-                                <span style={{ textDecoration: 'underline', cursor: 'pointer' }} onClick={() => this.handleLearnMore(0)}>
-                                    Learn more
-                                </span>
-                            </span>
-                        }
-                    >
-                        Select the research field
-                    </Tooltip>
-                </h2>
-                <CardDeck>
-                    {this.props.researchFields.length > 0 &&
-                        this.props.researchFields.map((fields, level) => {
-                            return fields && fields.length > 0 ? (
-                                <FieldSelector className="fieldSelector" key={level}>
-                                    <ListGroup flush>
-                                        <TransitionGroup exit={false}>
-                                            {fields.map(field => (
-                                                <ListGroupItemTransition key={field.id} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
-                                                    <ListGroupItemStyled
-                                                        style={{ display: 'flex' }}
-                                                        className="align-items-start"
-                                                        active={field.active}
-                                                        onClick={() => this.handleFieldClick(field.id, level)}
-                                                    >
-                                                        <div>{field.label}</div>
-                                                        {field.active &&
-                                                            level < this.props.researchFields.length - 1 &&
-                                                            this.props.researchFields[level + 1] &&
-                                                            this.props.researchFields[level + 1].length > 0 && (
-                                                                <Icon
-                                                                    size="1x"
-                                                                    style={{ marginLeft: 'auto' }}
-                                                                    className="flex-shrink-0  align-self-center"
-                                                                    icon={faAngleDoubleRight}
-                                                                />
-                                                            )}
-                                                    </ListGroupItemStyled>
-                                                </ListGroupItemTransition>
-                                            ))}
-                                        </TransitionGroup>
-                                    </ListGroup>
-                                </FieldSelector>
-                            ) : (
-                                ''
-                            );
-                        })}
-                </CardDeck>
-
-                {researchFieldLabel && this.props.selectedResearchField !== MISC.RESEARCH_FIELD_MAIN ? (
-                    <div className="mt-5 mb-3">
-                        Selected research field: <b>{researchFieldLabel}</b>
-                    </div>
-                ) : (
-                    <p className={errorMessageClasses} style={{ borderLeft: '4px red solid' }}>
-                        Please select the research field
-                    </p>
-                )}
-
-                <hr className="mt-5 mb-3" />
-                {/*<strong>Selected research field</strong> <br />
-                <span >{this.state.selectedResearchField}</span>*/}
-
-                <Button color="primary" className="float-right mb-4" onClick={this.handleNextClick}>
-                    Next step
-                </Button>
-                <Button color="light" className="float-right mb-4 mr-2" onClick={this.props.previousStep}>
-                    Previous step
-                </Button>
-                {this.props.researchFields.length > 0 && (
-                    <Tour
-                        onAfterOpen={this.disableBody}
-                        onBeforeClose={this.enableBody}
-                        disableInteraction={false}
-                        steps={[
-                            {
-                                selector: '.fieldSelector',
-                                content:
-                                    'Select a close research field to the paper from the list. The research field can be selected from a hierarchical structure of fields and their subfields.',
-                                style: { borderTop: '4px solid #E86161' }
-                            }
-                        ]}
-                        showNumber={false}
-                        accentColor={this.props.theme.orkgPrimaryColor}
-                        rounded={10}
-                        onRequestClose={this.requestCloseTour}
-                        isOpen={this.props.isTourOpen}
-                        startAt={this.props.tourStartAt}
-                        showButtons={false}
-                        showNavigation={false}
-                        maskClassName="reactourMask"
-                    />
-                )}
+    return (
+        <div>
+            <h2 className="h4 mt-4 mb-4">Select the research field</h2>
+            <p className="text-muted">
+                Select a research field by using the search field or select a field form the list {isLoading && <Icon icon={faSpinner} spin />}
+            </p>
+            <div className="mb-4">
+                <Autocomplete
+                    requestUrl={resourcesUrl}
+                    optionsClass={CLASSES.RESEARCH_FIELD}
+                    placeholder="Search for fields..."
+                    onItemSelected={handleFieldSelect}
+                    value={selectedResearchField !== MISC.RESEARCH_FIELD_MAIN ? { id: selectedResearchField, label: researchFieldLabel } : null}
+                    allowCreate={false}
+                    autoLoadOption={true}
+                />
             </div>
-        );
-    }
+
+            <CardDeck>
+                {researchFields.length > 0 &&
+                    researchFields.map((fields, level) => {
+                        return fields && fields.length > 0 ? (
+                            <FieldSelector className="fieldSelector" key={level}>
+                                <ListGroup flush>
+                                    <TransitionGroup exit={false}>
+                                        {fields.map(field => (
+                                            <ListGroupItemTransition key={field.id} classNames="fadeIn" timeout={{ enter: 500, exit: 0 }}>
+                                                <ListGroupItemStyled
+                                                    style={{ display: 'flex' }}
+                                                    className="align-items-start"
+                                                    active={field.active}
+                                                    onClick={() => handleFieldClick(field.id, level)}
+                                                >
+                                                    <div>{field.label}</div>
+                                                    {field.active &&
+                                                        level < researchFields.length - 1 &&
+                                                        researchFields[level + 1] &&
+                                                        researchFields[level + 1].length > 0 && (
+                                                            <Icon
+                                                                size="1x"
+                                                                style={{ marginLeft: 'auto' }}
+                                                                className="flex-shrink-0  align-self-center"
+                                                                icon={faAngleDoubleRight}
+                                                            />
+                                                        )}
+                                                </ListGroupItemStyled>
+                                            </ListGroupItemTransition>
+                                        ))}
+                                    </TransitionGroup>
+                                </ListGroup>
+                            </FieldSelector>
+                        ) : (
+                            ''
+                        );
+                    })}
+            </CardDeck>
+
+            {researchFieldLabel && selectedResearchField !== MISC.RESEARCH_FIELD_MAIN ? (
+                <div className="mt-5 mb-3">
+                    Selected research field: <b>{researchFieldLabel}</b>
+                </div>
+            ) : (
+                <p className={`text-danger mt-2 pl-2 ${!showError ? ' d-none' : ''}`} style={{ borderLeft: '4px red solid' }}>
+                    Please select the research field
+                </p>
+            )}
+
+            <hr className="mt-5 mb-3" />
+
+            <Button color="primary" className="float-right mb-4" onClick={handleNextClick}>
+                Next step
+            </Button>
+            <Button color="light" className="float-right mb-4 mr-2" onClick={() => dispatch(previousStep())}>
+                Previous step
+            </Button>
+        </div>
+    );
 }
 
-ResearchField.propTypes = {
-    selectedResearchField: PropTypes.string.isRequired,
-    researchFields: PropTypes.array.isRequired,
-    updateResearchField: PropTypes.func.isRequired,
-    nextStep: PropTypes.func.isRequired,
-    previousStep: PropTypes.func.isRequired,
-    theme: PropTypes.object.isRequired,
-    cookies: PropTypes.instanceOf(Cookies).isRequired,
-    openTour: PropTypes.func.isRequired,
-    closeTour: PropTypes.func.isRequired,
-    isTourOpen: PropTypes.bool.isRequired,
-    tourStartAt: PropTypes.number.isRequired
-};
-
-const mapStateToProps = state => ({
-    selectedResearchField: state.addPaper.selectedResearchField,
-    researchFields: state.addPaper.researchFields,
-    isTourOpen: state.addPaper.isTourOpen,
-    tourStartAt: state.addPaper.tourStartAt
-});
-
-const mapDispatchToProps = dispatch => ({
-    updateResearchField: data => dispatch(updateResearchField(data)),
-    nextStep: () => dispatch(nextStep()),
-    previousStep: () => dispatch(previousStep()),
-    openTour: data => dispatch(openTour(data)),
-    closeTour: () => dispatch(closeTour())
-});
-
-export default compose(
-    connect(
-        mapStateToProps,
-        mapDispatchToProps
-    ),
-    withTheme,
-    withCookies
-)(ResearchField);
+export default ResearchField;
