@@ -4,32 +4,35 @@ import styled from 'styled-components';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faFire, faSpinner } from '@fortawesome/free-solid-svg-icons';
 import { getTopResearchProblems } from 'services/backend/problems';
+import Tippy from '@tippy.js/react';
 import { reverse } from 'named-urls';
 import ROUTES from 'constants/routes';
 
-const List = styled.ul`
-    list-style: none;
+const List = styled.div`
     padding: 0 40px; // have a large horizontal padding to ensure the items are not out of the box on transform
-    li {
-        transition: transform 0.3s ease-in-out;
+    & > div {
         display: block;
 
-        &:hover {
+        div {
+            transition: transform 0.3s ease-in-out;
+        }
+
+        &:hover > div {
             transform: scale(1.2);
         }
         a {
             text-decoration: none;
         }
     }
-    li.item-0 {
-        font-size: 170%;
-        min-height: 45px;
+    .item-0 {
+        font-size: 145%;
+        min-height: 40px;
         a {
             color: ${props => props.theme.darkblueDarker}!important;
         }
     }
-    li.item-1,
-    li.item-2 {
+    .item-1,
+    .item-2 {
         font-size: 145%;
         min-height: 40px;
 
@@ -37,8 +40,8 @@ const List = styled.ul`
             color: ${props => props.theme.darkblue}!important;
         }
     }
-    li.item-3,
-    li.item-4 {
+    .item-3,
+    .item-4 {
         font-size: 100%;
         min-height: 30px;
 
@@ -72,8 +75,12 @@ const TrendingProblems = props => {
         <div>
             <div className="box rounded-lg mt-4">
                 <h2 className="h5 p-3 mb-0">
-                    <Icon icon={faFire} className="text-primary mr-2" />
-                    Trending research problems
+                    <Tippy content="A list of research problems addressed by papers recently added to ORKG.">
+                        <span>
+                            <Icon icon={faFire} className="text-primary mr-2" />
+                            Trending research problems
+                        </span>
+                    </Tippy>
                 </h2>
 
                 <hr className="mx-3 mt-0" />
@@ -83,12 +90,15 @@ const TrendingProblems = props => {
                         <Icon icon={faSpinner} spin /> Loading
                     </div>
                 ) : (
-                    <div className="px-4 text-center pb-1">
+                    <div className="px-4 text-center pb-5">
                         <List>
                             {problems.map((problem, index) => (
-                                <li className={`item-${index}`} key={`${index}-problem`}>
-                                    <Link to={reverse(ROUTES.RESEARCH_PROBLEM, { researchProblemId: problem.id })}>{problem.label}</Link>
-                                </li>
+                                <div key={`${index}-problem`}>
+                                    <div className={`item-${index}`}>
+                                        <Link to={reverse(ROUTES.RESEARCH_PROBLEM, { researchProblemId: problem.id })}>{problem.label}</Link>
+                                    </div>
+                                    {index !== problems.length - 1 && <hr className="mx-3 mb-1 mt-0" />}
+                                </div>
                             ))}
                         </List>
                     </div>
