@@ -1,14 +1,17 @@
 import React, { Component } from 'react';
-import { FormGroup, Label, Input, Button, Modal, ModalBody, ModalHeader, ModalFooter, FormFeedback } from 'reactstrap';
+import { FormGroup, Label, Button, Modal, ModalBody, ModalHeader, ModalFooter } from 'reactstrap';
 import { sortableContainer, sortableElement, sortableHandle } from 'react-sortable-hoc';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faTimes, faSpinner, faSort, faPen, faPlus } from '@fortawesome/free-solid-svg-icons';
 import { faOrcid } from '@fortawesome/free-brands-svg-icons';
 import styled, { withTheme } from 'styled-components';
-import { submitGetRequest } from 'network';
-import { get_error_message } from 'utils';
+import Autocomplete from 'components/Autocomplete/Autocomplete';
+import { resourcesUrl } from 'services/backend/resources';
+import { CLASSES } from 'constants/graphSettings';
+import { getPersonFullNameByORCID } from 'services/ORCID/index';
 import arrayMove from 'array-move';
 import PropTypes from 'prop-types';
+import { toast } from 'react-toastify';
 
 const AuthorTags = styled.div`
     display: flex;
@@ -113,8 +116,8 @@ class AuthorsInput extends Component {
         this.state = {
             showAuthorForm: false,
             authorInput: '',
+            authorAutocompleteLabel: '',
             authorNameLoading: false,
-            errors: null,
             editMode: false,
             editIndex: 0
         };
@@ -126,8 +129,11 @@ class AuthorsInput extends Component {
         }));
     };
 
-    handleChange = event => {
-        this.setState({ [event.target.name]: event.target.value });
+    handleChange = selected => {
+        if (selected.__isNew__) {
+            selected = { ...selected, label: selected.value };
+        }
+        this.setState({ authorInput: selected });
     };
 
     isORCID = value => {
@@ -137,25 +143,17 @@ class AuthorsInput extends Component {
         return Boolean(value && value.match(supportedORCID));
     };
 
-    getFullname = name => {
-        let fullName = name['family-name'] && name['family-name'].value ? name['family-name'].value : '';
-        fullName = name['given-names'] && name['given-names'].value ? `${name['given-names'].value} ${fullName}` : fullName;
-        return fullName.trim();
-    };
-
-    saveAuthor = () => {
-        if (this.state.authorInput) {
-            if (this.isORCID(this.state.authorInput)) {
+    saveAuthor = authorInput => {
+        if (authorInput && authorInput.label) {
+            if (this.isORCID(authorInput.label)) {
                 this.setState({ authorNameLoading: true });
                 // Get the full name from ORCID API
-                const orcid = this.state.authorInput.match(/([0-9]{4})-?([0-9]{4})-?([0-9]{4})-?(([0-9]{4})|([0-9]{3}X))/g)[0];
-                const ORCIDLink = 'https://pub.orcid.org/v2.0/' + orcid + '/person';
-                submitGetRequest(ORCIDLink, { Accept: 'application/orcid+json' })
-                    .then(response => {
-                        const authorName = this.getFullname(response.name);
+                const orcid = authorInput.label.match(/([0-9]{4})-?([0-9]{4})-?([0-9]{4})-?(([0-9]{4})|([0-9]{3}X))/g)[0];
+                getPersonFullNameByORCID(orcid)
+                    .then(authorFullName => {
                         const newAuthor = {
-                            label: authorName,
-                            id: authorName,
+                            label: authorFullName,
+                            id: authorFullName,
                             orcid: orcid,
                             statementId:
                                 this.state.editMode && this.props.value[this.state.editIndex] && this.props.value[this.state.editIndex].statementId
@@ -174,21 +172,21 @@ class AuthorsInput extends Component {
                         this.setState({
                             authorNameLoading: false,
                             authorInput: '',
-                            errors: null,
                             editMode: false
                         });
                         this.toggle('showAuthorForm');
                     })
                     .catch(e => {
                         this.setState({
-                            authorNameLoading: false,
-                            errors: { errors: [{ field: 'authorInput', message: `Invalid ORCID ID. Please enter the ${this.props.itemLabel} name` }] }
+                            authorNameLoading: false
                         });
+                        toast.error(`Invalid ORCID ID. Please enter the ${this.props.itemLabel} name`);
                     });
             } else {
                 const newAuthor = {
-                    label: this.state.authorInput,
-                    id: this.state.authorInput,
+                    ...authorInput,
+                    label: authorInput.label,
+                    id: authorInput.id ? authorInput.id : authorInput.label, // ID if the Author resource Exist
                     orcid: '',
                     statementId:
                         this.state.editMode && this.props.value[this.state.editIndex] && this.props.value[this.state.editIndex].statementId
@@ -204,11 +202,11 @@ class AuthorsInput extends Component {
                 } else {
                     this.props.handler([...this.props.value, newAuthor]);
                 }
-                this.setState({ authorInput: '', errors: null, editMode: false });
+                this.setState({ authorInput: '', editMode: false });
                 this.toggle('showAuthorForm');
             }
         } else {
-            this.setState({ errors: { errors: [{ field: 'authorInput', message: `Please enter the ${this.props.itemLabel} name` }] } });
+            toast.error(`Please enter the ${this.props.itemLabel} name`);
         }
     };
 
@@ -223,8 +221,7 @@ class AuthorsInput extends Component {
     editAuthor = key => {
         this.setState({
             editIndex: key,
-            authorInput: this.props.value[key].orcid ? this.props.value[key].orcid : this.props.value[key].label,
-            errors: null,
+            authorInput: this.props.value[key].orcid ? this.props.value[key].orcid : this.props.value[key],
             editMode: true
         });
         this.toggle('showAuthorForm');
@@ -269,7 +266,7 @@ class AuthorsInput extends Component {
                         color="light"
                         className="w-100"
                         onClick={() => {
-                            this.setState({ authorNameLoading: false, authorInput: '', errors: null, editMode: false });
+                            this.setState({ authorNameLoading: false, authorInput: '', editMode: false });
                             this.toggle('showAuthorForm');
                         }}
                     >
@@ -285,25 +282,31 @@ class AuthorsInput extends Component {
                             <Label for="authorInput">
                                 Enter {this.props.itemLabel} name <b>or</b> ORCID <Icon color="#A6CE39" icon={faOrcid} />
                             </Label>
-                            <Input
+                            <Autocomplete
+                                requestUrl={resourcesUrl}
+                                optionsClass={CLASSES.AUTHOR}
+                                placeholder="Search for author or enter a new author..."
                                 onChange={this.handleChange}
-                                type="text"
-                                name="authorInput"
-                                id="authorInput"
                                 value={this.state.authorInput}
-                                invalid={Boolean(get_error_message(this.state.errors, 'authorInput'))}
+                                allowCreate={true}
+                                autoLoadOption={false}
                                 innerRef={ref => (this.inputRef.current = ref)}
+                                inputId="authorInput"
+                                onChangeInputValue={value => this.setState({ authorAutocompleteLabel: value })}
                             />
-                            {Boolean(get_error_message(this.state.errors, 'authorInput')) && (
-                                <FormFeedback>{get_error_message(this.state.errors, 'authorInput')}</FormFeedback>
-                            )}
                         </FormGroup>
                     </ModalBody>
                     <ModalFooter>
                         <Button color="light" onClick={() => this.toggle('showAuthorForm')}>
                             Cancel
                         </Button>
-                        <Button disabled={this.state.authorNameLoading} color="primary" onClick={() => this.saveAuthor()}>
+                        <Button
+                            disabled={this.state.authorNameLoading}
+                            color="primary"
+                            onClick={() =>
+                                this.saveAuthor(this.state.authorInput ? this.state.authorInput : { label: this.state.authorAutocompleteLabel })
+                            }
+                        >
                             {!this.state.authorNameLoading ? this.state.editMode ? 'Save' : 'Add' : <Icon icon={faSpinner} spin />}
                         </Button>{' '}
                     </ModalFooter>
