@@ -149,7 +149,7 @@ class AddOrganization extends Component {
                         </Form>
                     ) : (
                         <>
-                            <Button color="link" className="p-0 mb-2 mt-2 clearfix" onClick={() => this.props.openAuthDialog('signin')}>
+                            <Button color="link" className="p-0 mb-2 mt-2 clearfix" onClick={() => this.props.openAuthDialog({ action: 'signin' })}>
                                 <Icon className="mr-1" icon={faUser} /> Signin to create organization
                             </Button>
                         </>
@@ -165,7 +165,7 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-    openAuthDialog: action => dispatch(openAuthDialog(action))
+    openAuthDialog: payload => dispatch(openAuthDialog(payload))
 });
 
 AddOrganization.propTypes = {
